# templates/nextjs-shadcn-ecommerce-admin/apps/api/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: "{{name}}"
  version: 1.0.0
  description: Minimal starter spec for this template's existing endpoints,
    generated from its Zod schemas (ADR 0040). Extend the schema/route modules
    this script imports as the project's real API contract grows — never
    hand-edit this file.
servers:
  - url: /api/v1
paths:
...
  /admin/products:
    get:
      operationId: listProducts
      summary: List products
      security:
        - sessionCookie: []
          apiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: size
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: A page of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      operationId: createProduct
      summary: Create a product
      security:
        - sessionCookie: []
          apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: The created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductOutput"
        "403":
          description: Not allowed to create a product
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: A product with this slug already exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /admin/products/{id}:
    get:
      operationId: getProductById
      summary: Get a product by id
      security:
        - sessionCookie: []
          apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductOutput"
        "404":
          description: Product not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      operationId: updateProduct
      summary: Update a product
      security:
        - sessionCookie: []
          apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: The updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductOutput"
        "403":
          description: Not allowed to update a product
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Product not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: A product with this slug already exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: deleteProduct
      summary: Delete (deactivate) a product
      security:
        - sessionCookie: []
          apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Product deactivated
        "403":
          description: Not allowed to delete a product
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Product not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
...
    ProductCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          example: Wireless Mouse
        code:
          example: WM-100
          type: string
        altCode:
          example: ALT-WM-100
          type: string
        description:
          example: Ergonomic wireless mouse.
          type: string
        price:
          type: number
          exclusiveMinimum: 0
          example: 29.99
        stock:
          example: 42
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        stockMin:
          example: 5
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        coverImage:
          example: https://images.example.com/wireless-mouse.jpg
          type: string
          format: uri
        categoryId:
          example: 9c4f3e1a-3b7e-4b1a-9c7a-4d3b6e2f8a1c
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      required:
        - name
        - price
      description: Creates a product. `slug` is derived from `name`.
    ProductUpdate:
      type: object
      properties:
        name:
          example: Wireless Mouse
          type: string
          minLength: 1
        code:
          example: WM-100
          anyOf:
            - type: string
            - type: "null"
        altCode:
          example: ALT-WM-100
          anyOf:
            - type: string
            - type: "null"
        description:
          example: Ergonomic wireless mouse.
          anyOf:
            - type: string
            - type: "null"
        price:
          example: 29.99
          type: number
          exclusiveMinimum: 0
        stock:
          example: 42
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        stockMin:
          example: 5
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        coverImage:
          example: https://images.example.com/wireless-mouse.jpg
          anyOf:
            - type: string
              format: uri
            - type: "null"
        categoryId:
          example: 9c4f3e1a-3b7e-4b1a-9c7a-4d3b6e2f8a1c
          anyOf:
            - type: string
              format: uuid
              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            - type: "null"
        isActive:
          example: true
          type: boolean
      description: Updates a product. Renaming re-derives `slug` from the new `name`.
...
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri-reference
          description: A URI reference identifying the problem type.
          example: https://coli.dev/errors/not-found
        status:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
          example: 404
        title:
          type: string
          example: Resource not found.
        detail:
          example: Resource not found. order 1
          type: string
        instance:
          example: /api/v1/orders/1
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: must be a valid email
            required:
              - field
              - message
            additionalProperties: false
      required:
        - type
        - status
        - title
      additionalProperties: false
...
    ProductList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductOutput"
        pagination:
          type: object
          properties:
            count:
              type: number
            next:
              type: number
            page:
              type: number
            previous:
              type: number
            size:
              type: number
            total:
              type: number
          required:
            - count
            - next
            - page
            - previous
            - size
            - total
          additionalProperties: false
      required:
        - items
        - pagination
      additionalProperties: false
      description: A page of products.
...
    ProductOutput:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          example: 9c4f3e1a-3b7e-4b1a-9c7a-4d3b6e2f8a1c
        name:
          type: string
          example: Wireless Mouse
        slug:
          type: string
          example: wireless-mouse
        code:
          anyOf:
            - type: string
            - type: "null"
          example: WM-100
        altCode:
          anyOf:
            - type: string
            - type: "null"
          example: ALT-WM-100
        description:
          anyOf:
            - type: string
            - type: "null"
          example: Ergonomic wireless mouse.
        price:
          type: number
          example: 29.99
        stock:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
          example: 42
        stockMin:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
          example: 5
        coverImage:
          anyOf:
            - type: string
              format: uri
            - type: "null"
          example: https://images.example.com/wireless-mouse.jpg
        categoryId:
          anyOf:
            - type: string
              format: uuid
              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            - type: "null"
          example: 9c4f3e1a-3b7e-4b1a-9c7a-4d3b6e2f8a1c
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
          example: 2026-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
          example: 2026-01-01T00:00:00Z
      required:
        - id
        - name
        - slug
        - code
        - altCode
        - description
        - price
        - stock
        - stockMin
        - coverImage
        - categoryId
        - isActive
        - createdAt
        - updatedAt
      additionalProperties: false
      description: A catalog product.
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-service-key